Re-jigg package structure to work with setuptools
