+++ .dockerignore
@@ -5,4 +5,4 @@
 .git
 .dockerignore
 __pycache__
-/test
+/toolbox/tests
