+++ .gitignore
@@ -1,7 +1,7 @@
 # Python temp file
 *.py[cod]
 __pycache__
-utilities/__pycache__
+toolbox/utilities/__pycache__
 .pytest_cache
 .python-version
 
@@ -26,7 +26,7 @@ lib64
 # Installer logs
 pip-log.txt
 
-# Unit test / coverage reports
+# Unit tests / coverage reports
 .coverage
 .coverage.*
 coverage.xml
@@ -61,9 +61,9 @@ venv
 
 node_modules/
 /static/
-tests/results
-tests/coverage
-tests/a11y
+toolbox/tests/results
+toolbox/tests/coverage
+toolbox/tests/a11y
 
 **/_sprite.scss
