+++ toolbox/bulk_processing/invalid_address_processor.py
@@ -1,10 +1,10 @@
 import uuid
 from datetime import datetime
 
-from bulk_processing.bulk_processor import BulkProcessor
-from bulk_processing.processor_interface import Processor
-from bulk_processing.validators import case_exists_by_id, is_uuid, max_length, mandatory
-from config import Config
+from toolbox.bulk_processing.bulk_processor import BulkProcessor
+from toolbox.bulk_processing.processor_interface import Processor
+from toolbox.bulk_processing.validators import case_exists_by_id, is_uuid, max_length, mandatory
+from toolbox.config import Config
 
 
 class InvalidAddressProcessor(Processor):
