+++ toolbox/questionnaire_linking.py
@@ -8,9 +8,9 @@ import requests
 from pika import BasicProperties
 from pika.spec import PERSISTENT_DELIVERY_MODE
 
-import qid_checksum_validator
-from config import Config
-from utilities.rabbit_context import RabbitContext
+from toolbox import qid_checksum_validator
+from toolbox.config import Config
+from toolbox.utilities.rabbit_context import RabbitContext
 
 CASE_REF_ERROR_COUNT = 0
 QID_ERROR_COUNT = 0
