+++ toolbox/reminder_batch_scheduler/reminder_batch.py
@@ -5,9 +5,9 @@ from datetime import datetime
 import rfc3339
 from termcolor import colored
 
-from config import Config
-from reminder_batch_scheduler import constants
-from utilities import db_helper
+from toolbox.config import Config
+from toolbox.reminder_batch_scheduler import constants
+from toolbox.utilities import db_helper
 
 
 def main(wave: int, starting_batch: int, max_cases: int, action_plan_id: uuid.UUID, insert_rules: bool = False,
